\name{APPLIANCE}
\alias{APPLIANCE}
\docType{data}
\title{
Appliance shipments
}
\description{
Appliance shipments from 1960 to 1985}
\usage{data("APPLIANCE")}
\format{
  A data frame with 26 observations on the following 7 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{Dishwasher}}{a numeric vector, Factory shipments (domestic) of dishwashers (thousands)}
    \item{\code{Disposal}}{a numeric vector, Factory shipments (domestic) of disposers (thousands)}
    \item{\code{Refrigerator}}{a numeric vector, Factory shipments (domestic) of refrigerators (thousands)}
    \item{\code{Washer}}{a numeric vector, Factory shipments (domestic) of washing machines (thousands)}
    \item{\code{DurableGoodsExp}}{a numeric vector, Durable goods expenditures (billions of 1972 dollars)}
    \item{\code{PrivateResInvest}}{a numeric vector, Private residential investment (billions of 1972 dollars)}
  }
}
\details{
From the Data and Story library \url{http://lib.stat.cmu.edu/DASL/Datafiles/appliancedat.html}

The file gives unit shipments of dishwashers, disposers, refrigerators, and washers in the United States from 1960 to 1985. This and other data are published currently in the Department of Commerce's Survey of Current Business, and are summarized from time to time in their publication, Business Statistics. Also included in the file are durable goods expenditures and private residential investment in the United States.
}
\keyword{datasets}
