\name{BODYFAT}
\alias{BODYFAT}
\docType{data}
\title{
BODYFAT data
}
\description{
Popular Bodyfat dataset 
}
\usage{data("BODYFAT")}
\format{
  A data frame with 252 observations on the following 14 variables.
  \describe{
    \item{\code{BodyFat}}{a numeric vector indicating the percentage body fat 0-100}
    \item{\code{Age}}{a numeric vector, yrs}
    \item{\code{Weight}}{a numeric vector, lbs}
    \item{\code{Height}}{a numeric vector, inches}
    \item{\code{Neck}}{a numeric vector}
    \item{\code{Chest}}{a numeric vector}
    \item{\code{Abdomen}}{a numeric vector}
    \item{\code{Hip}}{a numeric vector}
    \item{\code{Thigh}}{a numeric vector}
    \item{\code{Knee}}{a numeric vector}
    \item{\code{Ankle}}{a numeric vector}
    \item{\code{Biceps}}{a numeric vector}
    \item{\code{Forearm}}{a numeric vector}
    \item{\code{Wrist}}{a numeric vector}
  }
}
\details{
Bodyfat can be accurately measured by the hydrostatic technique, where someone is submereged in a tank of water.  It would be useful to be able to predict body fat from measurements that are simpler to obtain.  Unless otherwise specified, all physical measurements are in centimeters.
}
\source{
This is a modified version of the data available in ``Fitting Percentage of Body Fat to Simple Body Measurements" as appearing in Journal of Statistics Education v4 n1 (1996).  \url{http://www.amstat.org/publications/jse/v4n1/datasets.johnson.html}
}
\keyword{datasets}
