\name{BODYFAT2}
\alias{BODYFAT2}
\docType{data}
\title{
Secondary BODYFAT dataset
}
\description{
Bodyfat dataset illustrating quirks of statistical significance
}
\usage{data("BODYFAT2")}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{\code{Triceps}}{a numeric vector, cm}
    \item{\code{Thigh}}{a numeric vector, cm}
    \item{\code{Midarm}}{a numeric vector, cm}
    \item{\code{BodyFat}}{a numeric vector, 0-100 representing percent}
  }
}
\details{
The physical measurements are circumferences of body parts of 25-34 year-old healthy females.
}
\source{
This is a classic dataset found in many textbooks and in many places online.  The original source may be Neter, Kutner, Nachtsheim, Wasserman, 1997, p. 261: Applied Statistical Models (4th Edition). 
}

\keyword{datasets}
