\name{cor.demo}
\alias{cor.demo}

\title{
Correlation demo
}
\description{
This function shows the correlation and coefficient of determination as user interactively adds datapoints.  Useful for seeing what different values of correlation look like and seeing the effect of outliers.  }
\usage{
cor.demo()
}
\arguments{
No arguments
}
\details{
This function allows the user to generate data by click on a plot.  Once two points are added, the correlation (r) and coefficient of determination (r^2) are displayed.  When an additional point is added, these values are updated in the upper left with previous values being displayed in the upper right.  The effect of outliers on the correlation and coefficient of determination can easily be illustrated.  

Note:  To end the demo, you MUST click on the red box labeled "End" (or press Escape, which will return an error)
}

\author{
Adam Petrie
}

