\name{DIET}
\alias{DIET}
\docType{data}
\title{
DIET data
}
\description{
The weight of a person over time who is dieting and exercising
}
\usage{data("DIET")}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{Weight}}{a numeric vector, lbs}
    \item{\code{Day}}{a numeric vector, the number of days after the diet started}
  }
}
\details{
This data was collected by the author and consists of his weight measured first thing in the morning over the course of amount a month.  The scale round to the nearest 0.2 lbs.
}

\keyword{datasets}
