\name{EX2.TIPS}
\alias{EX2.TIPS}
\docType{data}
\title{
TIPS data for Exercise 6 in Chapter 2 
}
\description{
TIPS data for Exercise 6 in Chapter 2
}
\usage{data("EX2.TIPS")}
\format{
  A data frame with 244 observations on the following 8 variables.
  \describe{
    \item{\code{Tip.Percentage}}{a numeric vector}
    \item{\code{Bill_in_USD}}{a numeric vector}
    \item{\code{Tip_in_USD}}{a numeric vector}
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male}}
    \item{\code{Smoker}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{Weekday}}{a factor with levels \code{Friday} \code{Saturday} \code{Sunday} \code{Thursday}}
    \item{\code{Day_Night}}{a factor with levels \code{Day} \code{Night}}
    \item{\code{Size_of_Party}}{a numeric vector}
  }
}
\details{
See \code{TIPS} for more details.  This is the same dataset except that the names of the variables are different. 
}
\keyword{datasets}
