\name{EX3.BODYFAT}
\alias{EX3.BODYFAT}
\docType{data}
\title{
Bodyfat data for Exercise F in Chapter 3
}
\description{
Bodyfat data for Exercise F in Chapter 3
}
\usage{data("EX3.BODYFAT")}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{\code{Triceps}}{a numeric vector}
    \item{\code{Thigh}}{a numeric vector}
    \item{\code{Midarm}}{a numeric vector}
    \item{\code{Fat}}{a numeric vector}
  }
}
\details{
Same data as \code{BODYFAT2}, which you can see for more details.
}

\keyword{datasets}
