\name{EX3.HOUSING}
\alias{EX3.HOUSING}
\docType{data}
\title{
Housing data for Exercise E in Chapter 3}
\description{
Housing data for Exercise E in Chapter 3}

\usage{data("EX3.HOUSING")}
\format{
  A data frame with 522 observations on the following 2 variables.
  \describe{
    \item{\code{AREA}}{a numeric vector, square area of house}
    \item{\code{PRICE}}{a numeric vector, selling price}
  }
}
\details{
Selling prices of houses (perhaps in the Boston area in Massachusettes).
}
\source{
Original source unknown, but it appears in many places around the internet, e.g., \url{http://www.public.iastate.edu/~pdixon/stat500/data/realestate.txt} }
\keyword{datasets}
