\name{EX4.BIKE}
\alias{EX4.BIKE}
\docType{data}
\title{
Bike data for Exercise 1 in Chapter 4
}
\description{
Bike data for Exercise 1 in Chapter 4
}
\usage{data("EX4.BIKE")}
\format{
  A data frame with 414 observations on the following 5 variables.
  \describe{
    \item{\code{Demand}}{a numeric vector, total number of rental bikes}
    \item{\code{AvgTemp}}{a numeric vector, average temperature of the day}
    \item{\code{EffectiveAvgTemp}}{a numeric vector, average temperature it feels like (taking into account dewpoint) for the day}
    \item{\code{AvgHumidity}}{a numeric vector, average humidity for the day}
    \item{\code{AvgWindspeed}}{a numeric vector, average wind speed for the day}
  }
}
\details{
Adapted from the bike sharing dataset on the UCI data repository \url{http://archive.ics.uci.edu/ml/datasets/Bike+Sharing+Dataset}.  This concerns the demand for rental bikes in the DC area.

Bike sharing systems are new generation of traditional bike rentals where whole process from membership, rental and return back has become automatic. Through these systems, user is able to easily rent a bike from a particular position and return back at another position. Currently, there are about over 500 bike-sharing programs around the world which is composed of over 500 thousands bicycles. Today, there exists great interest in these systems due to their important role in traffic, environmental and health issues. 

Apart from interesting real world applications of bike sharing systems, the characteristics of data being generated by these systems make them attractive for the research. Opposed to other transport services such as bus or subway, the duration of travel, departure and arrival position is explicitly recorded in these systems. This feature turns bike sharing system into a virtual sensor network that can be used for sensing mobility in the city. Hence, it is expected that most of important events in the city could be detected via monitoring these data.
}
\references{
Fanaee-T, Hadi, and Gama, Joao, Event labeling combining ensemble detectors and background knowledge, Progress in Artificial Intelligence (2013): pp. 1-15, Springer Berlin Heidelberg.}
\keyword{datasets}
