\name{EX6.CLICK}
\alias{EX6.CLICK}
\docType{data}
\title{
CLICK data for Exercise 2 in Chapter 6}
\description{
CLICK data for Exercise 2 in Chapter 6}
\usage{data("EX6.CLICK")}
\format{
  A data frame with 13594 observations on the following 15 variables.
  \describe{
    \item{\code{Click}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{BannerPosition}}{a factor with levels \code{Pos1} \code{Pos2}, location of ad}
    \item{\code{SiteID}}{a factor with levels \code{S1} \code{S2} \code{S3} \code{S4} \code{S5} \code{S6} \code{S7} \code{S8}}
    \item{\code{SiteDomain}}{a factor with levels \code{SD1} \code{SD2} \code{SD3} \code{SD4} \code{SD5} \code{SD6} \code{SD7} \code{SD8}}
    \item{\code{SiteCategory}}{a factor with levels \code{SCat1} \code{SCat2} \code{SCat3} \code{SCat4} \code{SCat5}}
    \item{\code{AppDomain}}{a factor with levels \code{AD1} \code{AD2} \code{AD3}}
    \item{\code{AppCategory}}{a factor with levels \code{AC1} \code{AC2}}
    \item{\code{DeviceModel}}{a factor with levels \code{D1} \code{D10} \code{D11} \code{D12} \code{D13} \code{D14} \code{D15} \code{D16} \code{D17} \code{D18} \code{D2} \code{D3} \code{D4} \code{D5} \code{D6} \code{D7} \code{D8} \code{D9}}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E} \code{F} \code{G} \code{H} \code{I} \code{J} \code{K} \code{L} \code{M} \code{N} \code{O} \code{P} \code{Q} \code{R}}
    \item{\code{x3}}{a factor with levels \code{a} \code{b} \code{c} \code{d} \code{e} \code{f}}
    \item{\code{x4}}{a factor with levels \code{val1} \code{val2} \code{val3}}
    \item{\code{x5}}{a factor with levels \code{type1} \code{type2} \code{type3} \code{type4}}
    \item{\code{x6}}{a factor with levels \code{class1} \code{class2} \code{class3} \code{class4}}
    \item{\code{x7}}{a factor with levels \code{AA} \code{BB} \code{CC} \code{DD} \code{EE}}
  }
}
\details{
Inspired from a competition to predict the click-thru rates of ads displayed on mobile devices \url{https://www.kaggle.com/c/avazu-ctr-prediction}.  Does the click-thru rate vary based on where the ad placed, what kind of site and device is used to view the ad, something else?  All variables are anonymized.
}
\keyword{datasets}
