\name{EX6.DONOR}
\alias{EX6.DONOR}
\docType{data}
\title{
DONOR dataset for Exercise 1 in Chapter 6
}
\description{
DONOR dataset for Exercise 1 in Chapter 6
}
\usage{data("EX6.DONOR")}
\format{
  A data frame with 8132 observations on the following 18 variables.
  \describe{
    \item{\code{Donate}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{LastAmount}}{a numeric vector}
    \item{\code{AccountAge}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Setting}}{a factor with levels \code{Rural} \code{Suburban} \code{Urban}}
    \item{\code{Homeowner}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male} \code{Unknown}}
    \item{\code{Phone}}{a factor with levels \code{Listed} \code{Unlisted}}
    \item{\code{Source}}{a factor with levels \code{B} \code{M} \code{N} \code{P}}
    \item{\code{MedianHomeValue}}{a numeric vector}
    \item{\code{MedianIncome}}{a numeric vector}
    \item{\code{PercentOwnerOccupied}}{a numeric vector}
    \item{\code{Recent}}{a factor with levels \code{No} \code{Yes}}
    \item{\code{RecentResponsePercent}}{a numeric vector}
    \item{\code{RecentAvgAmount}}{a numeric vector}
    \item{\code{MonthsSinceLastGift}}{a numeric vector}
    \item{\code{TotalAmount}}{a numeric vector}
    \item{\code{TotalDonations}}{a numeric vector}
  }
}
\details{
Identical to \code{EX5.DONOR}, so see that for details}

\keyword{datasets}
