\name{EX7.BIKE}
\alias{EX7.BIKE}
\docType{data}
\title{
BIKE dataset for Exercise 1 Chapters 7 and 8}

\description{
BIKE dataset for Exercise 1 Chapters 7 and 8
}
\usage{data("EX7.BIKE")}
\format{
  A data frame with 410 observations on the following 9 variables.
  \describe{
    \item{\code{Demand}}{a numeric vector}
    \item{\code{Day}}{a factor with levels \code{Friday} \code{Monday} \code{Saturday} \code{Sunday} \code{Thursday} \code{Tuesday} \code{Wednesday}}
    \item{\code{Workingday}}{a factor with levels \code{no} \code{yes}}
    \item{\code{Holiday}}{a factor with levels \code{no} \code{yes}}
    \item{\code{Weather}}{a factor with levels \code{No rain} \code{Rain}}
    \item{\code{AvgTemp}}{a numeric vector}
    \item{\code{EffectiveAvgTemp}}{a numeric vector}
    \item{\code{AvgHumidity}}{a numeric vector}
    \item{\code{AvgWindspeed}}{a numeric vector}
  }
}
\details{
Identical to \code{EX5.BIKE} except with three additional rows deleted.  See that dataset for details.
}
\keyword{datasets}
