\name{EX9.BIRTHWEIGHT}
\alias{EX9.BIRTHWEIGHT}
\docType{data}
\title{
Birthweight dataset for Exercise 1 in Chapter 9
}
\description{
Birthweight dataset for Exercise 1 in Chapter 9
}
\usage{data("EX9.BIRTHWEIGHT")}
\format{
  A data frame with 553 observations on the following 13 variables.
  \describe{
    \item{\code{Birthweight}}{a numeric vector, grams}
    \item{\code{Gestation}}{a numeric vector, weeks}
    \item{\code{MotherRace}}{a factor with levels \code{Asian} \code{Black} \code{Mexican} \code{Mixed} \code{White}, self-reported}
    \item{\code{MotherAge}}{a numeric vector, self-reported}
    \item{\code{MotherEducation}}{a factor with levels \code{below HS} \code{College} \code{HS}, self-reported}
    \item{\code{MotherHeight}}{a numeric vector, inches}
    \item{\code{MotherWeight}}{a numeric vector, pounds}
    \item{\code{FatherRace}}{a factor with levels \code{Asian} \code{Black} \code{Mexican} \code{Mixed} \code{White}, self-reported}
    \item{\code{FatherAge}}{a numeric vector, self-reported}
    \item{\code{Father_Education}}{a factor with levels \code{below HS} \code{College} \code{HS}, self-reported}
    \item{\code{FatherHeight}}{a numeric vector, inches}
    \item{\code{FatherWeight}}{a numeric vector, pounds}
    \item{\code{Smoking}}{a factor with levels \code{never} \code{now}, self-reported}
  }
}
\details{
An examination of birthweights and their link to gestation, mother and father characteristics, and whether the mother smoked during pregnancy.  
}
\source{
	Adapted from a subset of a study from Nolan and Speed (2000) consisting of male, single births which survived for at least 28 days.  Some rows that contained bad data have been omitted.  \url{http://had.co.nz/stat645/week-05/birthweight.txt}}
\keyword{datasets}
