\name{EX9.NFL}
\alias{EX9.NFL}
\docType{data}
\title{
NFL data for Exercise 2 Chapter 9
}
\description{
NFL data for Exercise 2 Chapter 9
}
\usage{data("EX9.NFL")}
\format{
  A data frame with 352 observations on the following 26 variables.
  \describe{
    \item{\code{Wins}}{a numeric vector}
    \item{\code{X1.OffTotPlays}}{a numeric vector}
    \item{\code{X2.OffTotYdsperPly}}{a numeric vector}
    \item{\code{X3.OffPass1stDwns}}{a numeric vector}
    \item{\code{X4.OffRush1stDwns}}{a numeric vector}
    \item{\code{X5.OffFumblesLost}}{a numeric vector}
    \item{\code{X6.OffPassComp}}{a numeric vector}
    \item{\code{X7.OffPassINT}}{a numeric vector}
    \item{\code{X8.OffPassLongest}}{a numeric vector}
    \item{\code{X9.OffPassYdsperAtt}}{a numeric vector}
    \item{\code{X10.OffPassYdsperComp}}{a numeric vector}
    \item{\code{X11.OffPassSackYds}}{a numeric vector}
    \item{\code{X12.OffPassSack}}{a numeric vector}
    \item{\code{X13.OffRushLongest}}{a numeric vector}
    \item{\code{X14.OffRushYdsperAtt}}{a numeric vector}
    \item{\code{X15.OffRushYdsperGame}}{a numeric vector}
    \item{\code{X16.OffFumbles}}{a numeric vector}
    \item{\code{X17.1to29ydFG}}{a numeric vector}
    \item{\code{X18.30to39ydFG}}{a numeric vector}
    \item{\code{X19.40.ydFG}}{a numeric vector}
    \item{\code{X20.TotalFGAtt}}{a numeric vector}
    \item{\code{X21.OffTimesPunted}}{a numeric vector}
    \item{\code{X22.OffTimesHadPuntBlocked}}{a numeric vector}
    \item{\code{X23.OffYardsPerPunt}}{a numeric vector}
    \item{\code{X24.Off2ptConvMade}}{a numeric vector}
    \item{\code{X25.OffSafeties}}{a numeric vector}
  }
}
\details{
A subset of the \code{NFL} data (see entry for details) containing statistics on the offense.
}
\keyword{datasets}
