\name{EX9.STORE}
\alias{EX9.STORE}
\docType{data}
\title{
Data for Exercise 3 Chapter 9
}
\description{
Data for Exercise 3 Chapter 9
}
\usage{data("EX9.STORE")}
\format{
  A data frame with 1500 observations on the following 68 variables.
  \describe{
    \item{\code{Store1}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store2}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store3}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store4}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store5}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store6}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store7}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store8}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store9}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store10}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store11}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store12}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store13}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store14}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store15}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store16}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store17}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store18}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store19}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store20}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store21}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store22}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store23}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store24}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store25}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store26}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store27}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store28}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store29}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store30}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store31}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store32}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store33}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store34}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store35}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store36}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store37}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store38}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store39}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store40}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store41}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store42}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store43}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store44}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store45}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store46}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store47}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store48}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store49}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store50}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store51}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store52}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store53}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store54}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store55}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store56}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store57}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store58}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store59}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store60}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store61}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store62}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store63}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store64}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store65}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store66}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store67}}{a factor with levels \code{Buy} \code{No}}
    \item{\code{Store68}}{a factor with levels \code{Buy} \code{No}}
  }
}
\details{
The data consists of a random sample of 1500 credit card customers and their shopping habits regarding 68 different stores (whether they did or did not make a purchase in the last 90 days).  Shoppers don't pick and choose places to shop at random, so it is interesting to study which stores appear together in a customers' history.
}
\source{
Consultation with an anonymous client.  Stores have been anonymized to protect the source.
}
\keyword{datasets}
