\name{FRIEND}
\alias{FRIEND}
\docType{data}
\title{
Friendship Potential vs. Attractiveness Ratings
}
\description{
Examining the relationship between how likely someone would be friends with a person based on that person's level of attractiveness
}
\usage{data("FRIEND")}
\format{
  A data frame with 54 observations on the following 2 variables.
  \describe{
    \item{\code{Attractiveness}}{a numeric vector - the average scores (1-5) from about 80 male students who rated the attractiveness of the women in each picture }
    \item{\code{FriendshipPotential}}{a numeric vector - the average scores (1-5) from about 30 female students who rated how likely they would be to be friends with the pictured woman }
  }
}
\details{
The data contain information on 54 pictures of women posted on the (now defunct/renamed) site hotornot.com.  The women in two classes of introductory statistics at the University of Tennessee rated how likely they would be friends with the pictured women (on a scale of 1-5, 1 being very unlikely and 5 being very likely).  The men in three (different) classes of introductory statistics gave an attractiveness score to each woman (on a scale of 1-5, 1 being very unattractive and 5 being very attractive).  The numbers presented are the averages over all student ratings.  
}
\source{
Surveys administered to introductory statistics students at the University of Tennessee from 2008-2010.
}

\keyword{datasets}
