\name{FUMBLES}
\alias{FUMBLES}
\docType{data}
\title{
Wins vs. Fumbles of an NFL team
}
\description{
Wins vs. Fumbles of an NFL team}
\usage{data("FUMBLES")}
\format{
  A data frame with 352 observations on the following 2 variables.
  \describe{
    \item{\code{Wins}}{a numeric vector, number of wins (0-16) of an NFL team over the course of a season}
    \item{\code{FumblesLost}}{a numeric vector, the number of fumbles lost by that team over the course of a season}
  }
}
\details{
This is a subset of the \code{NFL} data.  Data is from the 2002-2012 seasons.
}
\source{
Collected by an undergraduate student from available web data in 2013.
}
\keyword{datasets}
