\name{LARGEFLYER}
\alias{LARGEFLYER}
\docType{data}
\title{
Interest in frequent flier program (large version)
}
\description{
Interest in frequent flier program (artificial)
}
\usage{data("LARGEFLYER")}
\format{
  A data frame with 100000 observations on the following 2 variables.
  \describe{
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male}}
    \item{\code{Interest}}{a factor with levels \code{No} \code{Yes}}
  }
}
\details{
This artificial datasets tabulates the interest in a new frequent flyer program based on gender.  It illustrates that a statistically significant association may have absolutely no practical significance.
}
\keyword{datasets}
