\name{LAUNCH}
\alias{LAUNCH}
\docType{data}
\title{
New product launch data
}
\description{
The profit of newly released products over the first few months of their release
}
\usage{data("LAUNCH")}
\format{
  A data frame with 652 observations on the following 420 variables.
  \describe{
    \item{\code{Profit}}{an anonymized numeric vector, the profit from the product over the first few months of release}
    \item{\code{x1}}{an anonymized numeric vector}
    \item{\code{x2}}{an anonymized numeric vector}
    \item{\code{x3}}{an anonymized numeric vector}
    \item{\code{x4}}{an anonymized numeric vector}
    \item{\code{x5}}{an anonymized numeric vector}
    \item{\code{x6}}{an anonymized numeric vector}
    \item{\code{x7}}{an anonymized numeric vector}
    \item{\code{x8}}{an anonymized numeric vector}
    \item{\code{x9}}{an anonymized numeric vector}
    \item{\code{x10}}{an anonymized numeric vector}
    \item{\code{x11}}{an anonymized numeric vector}
    \item{\code{x12}}{an anonymized numeric vector}
    \item{\code{x13}}{an anonymized numeric vector}
    \item{\code{x14}}{an anonymized numeric vector}
    \item{\code{x15}}{an anonymized numeric vector}
    \item{\code{x16}}{an anonymized numeric vector}
    \item{\code{x17}}{an anonymized numeric vector}
    \item{\code{x18}}{an anonymized numeric vector}
    \item{\code{x19}}{an anonymized numeric vector}
    \item{\code{x20}}{an anonymized numeric vector}
    \item{\code{x21}}{an anonymized numeric vector}
    \item{\code{x22}}{an anonymized numeric vector}
    \item{\code{x23}}{an anonymized numeric vector}
    \item{\code{x24}}{an anonymized numeric vector}
    \item{\code{x25}}{an anonymized numeric vector}
    \item{\code{x26}}{an anonymized numeric vector}
    \item{\code{x27}}{an anonymized numeric vector}
    \item{\code{x28}}{an anonymized numeric vector}
    \item{\code{x29}}{an anonymized numeric vector}
    \item{\code{x30}}{an anonymized numeric vector}
    \item{\code{x31}}{an anonymized numeric vector}
    \item{\code{x32}}{an anonymized numeric vector}
    \item{\code{x33}}{an anonymized numeric vector}
    \item{\code{x34}}{an anonymized numeric vector}
    \item{\code{x35}}{an anonymized numeric vector}
    \item{\code{x36}}{an anonymized numeric vector}
    \item{\code{x37}}{an anonymized numeric vector}
    \item{\code{x38}}{an anonymized numeric vector}
    \item{\code{x39}}{an anonymized numeric vector}
    \item{\code{x40}}{an anonymized numeric vector}
    \item{\code{x41}}{an anonymized numeric vector}
    \item{\code{x42}}{an anonymized numeric vector}
    \item{\code{x43}}{an anonymized numeric vector}
    \item{\code{x44}}{an anonymized numeric vector}
    \item{\code{x45}}{an anonymized numeric vector}
    \item{\code{x46}}{an anonymized numeric vector}
    \item{\code{x47}}{an anonymized numeric vector}
    \item{\code{x48}}{an anonymized numeric vector}
    \item{\code{x49}}{an anonymized numeric vector}
    \item{\code{x50}}{an anonymized numeric vector}
    \item{\code{x51}}{an anonymized numeric vector}
    \item{\code{x52}}{an anonymized numeric vector}
    \item{\code{x53}}{an anonymized numeric vector}
    \item{\code{x54}}{an anonymized numeric vector}
    \item{\code{x55}}{an anonymized numeric vector}
    \item{\code{x56}}{an anonymized numeric vector}
    \item{\code{x57}}{an anonymized numeric vector}
    \item{\code{x58}}{an anonymized numeric vector}
    \item{\code{x59}}{an anonymized numeric vector}
    \item{\code{x60}}{an anonymized numeric vector}
    \item{\code{x61}}{an anonymized numeric vector}
    \item{\code{x62}}{an anonymized numeric vector}
    \item{\code{x63}}{an anonymized numeric vector}
    \item{\code{x64}}{an anonymized numeric vector}
    \item{\code{x65}}{an anonymized numeric vector}
    \item{\code{x66}}{an anonymized numeric vector}
    \item{\code{x67}}{an anonymized numeric vector}
    \item{\code{x68}}{an anonymized numeric vector}
    \item{\code{x69}}{an anonymized numeric vector}
    \item{\code{x70}}{an anonymized numeric vector}
    \item{\code{x71}}{an anonymized numeric vector}
    \item{\code{x72}}{an anonymized numeric vector}
    \item{\code{x73}}{an anonymized numeric vector}
    \item{\code{x74}}{an anonymized numeric vector}
    \item{\code{x75}}{an anonymized numeric vector}
    \item{\code{x76}}{an anonymized numeric vector}
    \item{\code{x77}}{an anonymized numeric vector}
    \item{\code{x78}}{an anonymized numeric vector}
    \item{\code{x79}}{an anonymized numeric vector}
    \item{\code{x80}}{an anonymized numeric vector}
    \item{\code{x81}}{an anonymized numeric vector}
    \item{\code{x82}}{an anonymized numeric vector}
    \item{\code{x83}}{an anonymized numeric vector}
    \item{\code{x84}}{an anonymized numeric vector}
    \item{\code{x85}}{an anonymized numeric vector}
    \item{\code{x86}}{an anonymized numeric vector}
    \item{\code{x87}}{an anonymized numeric vector}
    \item{\code{x88}}{an anonymized numeric vector}
    \item{\code{x89}}{an anonymized numeric vector}
    \item{\code{x90}}{an anonymized numeric vector}
    \item{\code{x91}}{an anonymized numeric vector}
    \item{\code{x92}}{an anonymized numeric vector}
    \item{\code{x93}}{an anonymized numeric vector}
    \item{\code{x94}}{an anonymized numeric vector}
    \item{\code{x95}}{an anonymized numeric vector}
    \item{\code{x96}}{an anonymized numeric vector}
    \item{\code{x97}}{an anonymized numeric vector}
    \item{\code{x98}}{an anonymized numeric vector}
    \item{\code{x99}}{an anonymized numeric vector}
    \item{\code{x100}}{an anonymized numeric vector}
    \item{\code{x101}}{an anonymized numeric vector}
    \item{\code{x102}}{an anonymized numeric vector}
    \item{\code{x103}}{an anonymized numeric vector}
    \item{\code{x104}}{an anonymized numeric vector}
    \item{\code{x105}}{an anonymized numeric vector}
    \item{\code{x106}}{an anonymized numeric vector}
    \item{\code{x107}}{an anonymized numeric vector}
    \item{\code{x108}}{an anonymized numeric vector}
    \item{\code{x109}}{an anonymized numeric vector}
    \item{\code{x110}}{an anonymized numeric vector}
    \item{\code{x111}}{an anonymized numeric vector}
    \item{\code{x112}}{an anonymized numeric vector}
    \item{\code{x113}}{an anonymized numeric vector}
    \item{\code{x114}}{an anonymized numeric vector}
    \item{\code{x115}}{an anonymized numeric vector}
    \item{\code{x116}}{an anonymized numeric vector}
    \item{\code{x117}}{an anonymized numeric vector}
    \item{\code{x118}}{an anonymized numeric vector}
    \item{\code{x119}}{an anonymized numeric vector}
    \item{\code{x120}}{an anonymized numeric vector}
    \item{\code{x121}}{an anonymized numeric vector}
    \item{\code{x122}}{an anonymized numeric vector}
    \item{\code{x123}}{an anonymized numeric vector}
    \item{\code{x124}}{an anonymized numeric vector}
    \item{\code{x125}}{an anonymized numeric vector}
    \item{\code{x126}}{an anonymized numeric vector}
    \item{\code{x127}}{an anonymized numeric vector}
    \item{\code{x128}}{an anonymized numeric vector}
    \item{\code{x129}}{an anonymized numeric vector}
    \item{\code{x130}}{an anonymized numeric vector}
    \item{\code{x131}}{an anonymized numeric vector}
    \item{\code{x132}}{an anonymized numeric vector}
    \item{\code{x133}}{an anonymized numeric vector}
    \item{\code{x134}}{an anonymized numeric vector}
    \item{\code{x135}}{an anonymized numeric vector}
    \item{\code{x136}}{an anonymized numeric vector}
    \item{\code{x137}}{an anonymized numeric vector}
    \item{\code{x138}}{an anonymized numeric vector}
    \item{\code{x139}}{an anonymized numeric vector}
    \item{\code{x140}}{an anonymized numeric vector}
    \item{\code{x141}}{an anonymized numeric vector}
    \item{\code{x142}}{an anonymized numeric vector}
    \item{\code{x143}}{an anonymized numeric vector}
    \item{\code{x144}}{an anonymized numeric vector}
    \item{\code{x145}}{an anonymized numeric vector}
    \item{\code{x146}}{an anonymized numeric vector}
    \item{\code{x147}}{an anonymized numeric vector}
    \item{\code{x148}}{an anonymized numeric vector}
    \item{\code{x149}}{an anonymized numeric vector}
    \item{\code{x150}}{an anonymized numeric vector}
    \item{\code{x151}}{an anonymized numeric vector}
    \item{\code{x152}}{an anonymized numeric vector}
    \item{\code{x153}}{an anonymized numeric vector}
    \item{\code{x154}}{an anonymized numeric vector}
    \item{\code{x155}}{an anonymized numeric vector}
    \item{\code{x156}}{an anonymized numeric vector}
    \item{\code{x157}}{an anonymized numeric vector}
    \item{\code{x158}}{an anonymized numeric vector}
    \item{\code{x159}}{an anonymized numeric vector}
    \item{\code{x160}}{an anonymized numeric vector}
    \item{\code{x161}}{an anonymized numeric vector}
    \item{\code{x162}}{an anonymized numeric vector}
    \item{\code{x163}}{an anonymized numeric vector}
    \item{\code{x164}}{an anonymized numeric vector}
    \item{\code{x165}}{an anonymized numeric vector}
    \item{\code{x166}}{an anonymized numeric vector}
    \item{\code{x167}}{an anonymized numeric vector}
    \item{\code{x168}}{an anonymized numeric vector}
    \item{\code{x169}}{an anonymized numeric vector}
    \item{\code{x170}}{an anonymized numeric vector}
    \item{\code{x171}}{an anonymized numeric vector}
    \item{\code{x172}}{an anonymized numeric vector}
    \item{\code{x173}}{an anonymized numeric vector}
    \item{\code{x174}}{an anonymized numeric vector}
    \item{\code{x175}}{an anonymized numeric vector}
    \item{\code{x176}}{an anonymized numeric vector}
    \item{\code{x177}}{an anonymized numeric vector}
    \item{\code{x178}}{an anonymized numeric vector}
    \item{\code{x179}}{an anonymized numeric vector}
    \item{\code{x180}}{an anonymized numeric vector}
    \item{\code{x181}}{an anonymized numeric vector}
    \item{\code{x182}}{an anonymized numeric vector}
    \item{\code{x183}}{an anonymized numeric vector}
    \item{\code{x184}}{an anonymized numeric vector}
    \item{\code{x185}}{an anonymized numeric vector}
    \item{\code{x186}}{an anonymized numeric vector}
    \item{\code{x187}}{an anonymized numeric vector}
    \item{\code{x188}}{an anonymized numeric vector}
    \item{\code{x189}}{an anonymized numeric vector}
    \item{\code{x190}}{an anonymized numeric vector}
    \item{\code{x191}}{an anonymized numeric vector}
    \item{\code{x192}}{an anonymized numeric vector}
    \item{\code{x193}}{an anonymized numeric vector}
    \item{\code{x194}}{an anonymized numeric vector}
    \item{\code{x195}}{an anonymized numeric vector}
    \item{\code{x196}}{an anonymized numeric vector}
    \item{\code{x197}}{an anonymized numeric vector}
    \item{\code{x198}}{an anonymized numeric vector}
    \item{\code{x199}}{an anonymized numeric vector}
    \item{\code{x200}}{an anonymized numeric vector}
    \item{\code{x201}}{an anonymized numeric vector}
    \item{\code{x202}}{an anonymized numeric vector}
    \item{\code{x203}}{an anonymized numeric vector}
    \item{\code{x204}}{an anonymized numeric vector}
    \item{\code{x205}}{an anonymized numeric vector}
    \item{\code{x206}}{an anonymized numeric vector}
    \item{\code{x207}}{an anonymized numeric vector}
    \item{\code{x208}}{an anonymized numeric vector}
    \item{\code{x209}}{an anonymized numeric vector}
    \item{\code{x210}}{an anonymized numeric vector}
    \item{\code{x211}}{an anonymized numeric vector}
    \item{\code{x212}}{an anonymized numeric vector}
    \item{\code{x213}}{an anonymized numeric vector}
    \item{\code{x214}}{an anonymized numeric vector}
    \item{\code{x215}}{an anonymized numeric vector}
    \item{\code{x216}}{an anonymized numeric vector}
    \item{\code{x217}}{an anonymized numeric vector}
    \item{\code{x218}}{an anonymized numeric vector}
    \item{\code{x219}}{an anonymized numeric vector}
    \item{\code{x220}}{an anonymized numeric vector}
    \item{\code{x221}}{an anonymized numeric vector}
    \item{\code{x222}}{an anonymized numeric vector}
    \item{\code{x223}}{an anonymized numeric vector}
    \item{\code{x224}}{an anonymized numeric vector}
    \item{\code{x225}}{an anonymized numeric vector}
    \item{\code{x226}}{an anonymized numeric vector}
    \item{\code{x227}}{an anonymized numeric vector}
    \item{\code{x228}}{an anonymized numeric vector}
    \item{\code{x229}}{an anonymized numeric vector}
    \item{\code{x230}}{an anonymized numeric vector}
    \item{\code{x231}}{an anonymized numeric vector}
    \item{\code{x232}}{an anonymized numeric vector}
    \item{\code{x233}}{an anonymized numeric vector}
    \item{\code{x234}}{an anonymized numeric vector}
    \item{\code{x235}}{an anonymized numeric vector}
    \item{\code{x236}}{an anonymized numeric vector}
    \item{\code{x237}}{an anonymized numeric vector}
    \item{\code{x238}}{an anonymized numeric vector}
    \item{\code{x239}}{an anonymized numeric vector}
    \item{\code{x240}}{an anonymized numeric vector}
    \item{\code{x241}}{an anonymized numeric vector}
    \item{\code{x242}}{an anonymized numeric vector}
    \item{\code{x243}}{an anonymized numeric vector}
    \item{\code{x244}}{an anonymized numeric vector}
    \item{\code{x245}}{an anonymized numeric vector}
    \item{\code{x246}}{an anonymized numeric vector}
    \item{\code{x247}}{an anonymized numeric vector}
    \item{\code{x248}}{an anonymized numeric vector}
    \item{\code{x249}}{an anonymized numeric vector}
    \item{\code{x250}}{an anonymized numeric vector}
    \item{\code{x251}}{an anonymized numeric vector}
    \item{\code{x252}}{an anonymized numeric vector}
    \item{\code{x253}}{an anonymized numeric vector}
    \item{\code{x254}}{an anonymized numeric vector}
    \item{\code{x255}}{an anonymized numeric vector}
    \item{\code{x256}}{an anonymized numeric vector}
    \item{\code{x257}}{an anonymized numeric vector}
    \item{\code{x258}}{an anonymized numeric vector}
    \item{\code{x259}}{an anonymized numeric vector}
    \item{\code{x260}}{an anonymized numeric vector}
    \item{\code{x261}}{an anonymized numeric vector}
    \item{\code{x262}}{an anonymized numeric vector}
    \item{\code{x263}}{an anonymized numeric vector}
    \item{\code{x264}}{an anonymized numeric vector}
    \item{\code{x265}}{an anonymized numeric vector}
    \item{\code{x266}}{an anonymized numeric vector}
    \item{\code{x267}}{an anonymized numeric vector}
    \item{\code{x268}}{an anonymized numeric vector}
    \item{\code{x269}}{an anonymized numeric vector}
    \item{\code{x270}}{an anonymized numeric vector}
    \item{\code{x271}}{an anonymized numeric vector}
    \item{\code{x272}}{an anonymized numeric vector}
    \item{\code{x273}}{an anonymized numeric vector}
    \item{\code{x274}}{an anonymized numeric vector}
    \item{\code{x275}}{an anonymized numeric vector}
    \item{\code{x276}}{an anonymized numeric vector}
    \item{\code{x277}}{an anonymized numeric vector}
    \item{\code{x278}}{an anonymized numeric vector}
    \item{\code{x279}}{an anonymized numeric vector}
    \item{\code{x280}}{an anonymized numeric vector}
    \item{\code{x281}}{an anonymized numeric vector}
    \item{\code{x282}}{an anonymized numeric vector}
    \item{\code{x283}}{an anonymized numeric vector}
    \item{\code{x284}}{an anonymized numeric vector}
    \item{\code{x285}}{an anonymized numeric vector}
    \item{\code{x286}}{an anonymized numeric vector}
    \item{\code{x287}}{an anonymized numeric vector}
    \item{\code{x288}}{an anonymized numeric vector}
    \item{\code{x289}}{an anonymized numeric vector}
    \item{\code{x290}}{an anonymized numeric vector}
    \item{\code{x291}}{an anonymized numeric vector}
    \item{\code{x292}}{an anonymized numeric vector}
    \item{\code{x293}}{an anonymized numeric vector}
    \item{\code{x294}}{an anonymized numeric vector}
    \item{\code{x295}}{an anonymized numeric vector}
    \item{\code{x296}}{an anonymized numeric vector}
    \item{\code{x297}}{an anonymized numeric vector}
    \item{\code{x298}}{an anonymized numeric vector}
    \item{\code{x299}}{an anonymized numeric vector}
    \item{\code{x300}}{an anonymized numeric vector}
    \item{\code{x301}}{an anonymized numeric vector}
    \item{\code{x302}}{an anonymized numeric vector}
    \item{\code{x303}}{an anonymized numeric vector}
    \item{\code{x304}}{an anonymized numeric vector}
    \item{\code{x305}}{an anonymized numeric vector}
    \item{\code{x306}}{an anonymized numeric vector}
    \item{\code{x307}}{an anonymized numeric vector}
    \item{\code{x308}}{an anonymized numeric vector}
    \item{\code{x309}}{an anonymized numeric vector}
    \item{\code{x310}}{an anonymized numeric vector}
    \item{\code{x311}}{an anonymized numeric vector}
    \item{\code{x312}}{an anonymized numeric vector}
    \item{\code{x313}}{an anonymized numeric vector}
    \item{\code{x314}}{an anonymized numeric vector}
    \item{\code{x315}}{an anonymized numeric vector}
    \item{\code{x316}}{an anonymized numeric vector}
    \item{\code{x317}}{an anonymized numeric vector}
    \item{\code{x318}}{an anonymized numeric vector}
    \item{\code{x319}}{an anonymized numeric vector}
    \item{\code{x320}}{an anonymized numeric vector}
    \item{\code{x321}}{an anonymized numeric vector}
    \item{\code{x322}}{an anonymized numeric vector}
    \item{\code{x323}}{an anonymized numeric vector}
    \item{\code{x324}}{an anonymized numeric vector}
    \item{\code{x325}}{an anonymized numeric vector}
    \item{\code{x326}}{an anonymized numeric vector}
    \item{\code{x327}}{an anonymized numeric vector}
    \item{\code{x328}}{an anonymized numeric vector}
    \item{\code{x329}}{an anonymized numeric vector}
    \item{\code{x330}}{an anonymized numeric vector}
    \item{\code{x331}}{an anonymized numeric vector}
    \item{\code{x332}}{an anonymized numeric vector}
    \item{\code{x333}}{an anonymized numeric vector}
    \item{\code{x334}}{an anonymized numeric vector}
    \item{\code{x335}}{an anonymized numeric vector}
    \item{\code{x336}}{an anonymized numeric vector}
    \item{\code{x337}}{an anonymized numeric vector}
    \item{\code{x338}}{an anonymized numeric vector}
    \item{\code{x339}}{an anonymized numeric vector}
    \item{\code{x340}}{an anonymized numeric vector}
    \item{\code{x341}}{an anonymized numeric vector}
    \item{\code{x342}}{an anonymized numeric vector}
    \item{\code{x343}}{an anonymized numeric vector}
    \item{\code{x344}}{an anonymized numeric vector}
    \item{\code{x345}}{an anonymized numeric vector}
    \item{\code{x346}}{an anonymized numeric vector}
    \item{\code{x347}}{an anonymized numeric vector}
    \item{\code{x348}}{an anonymized numeric vector}
    \item{\code{x349}}{an anonymized numeric vector}
    \item{\code{x350}}{an anonymized numeric vector}
    \item{\code{x351}}{an anonymized numeric vector}
    \item{\code{x352}}{an anonymized numeric vector}
    \item{\code{x353}}{an anonymized numeric vector}
    \item{\code{x354}}{an anonymized numeric vector}
    \item{\code{x355}}{an anonymized numeric vector}
    \item{\code{x356}}{an anonymized numeric vector}
    \item{\code{x357}}{an anonymized numeric vector}
    \item{\code{x358}}{an anonymized numeric vector}
    \item{\code{x359}}{an anonymized numeric vector}
    \item{\code{x360}}{an anonymized numeric vector}
    \item{\code{x361}}{an anonymized numeric vector}
    \item{\code{x362}}{an anonymized numeric vector}
    \item{\code{x363}}{an anonymized numeric vector}
    \item{\code{x364}}{an anonymized numeric vector}
    \item{\code{x365}}{an anonymized numeric vector}
    \item{\code{x366}}{an anonymized numeric vector}
    \item{\code{x367}}{an anonymized numeric vector}
    \item{\code{x368}}{an anonymized numeric vector}
    \item{\code{x369}}{an anonymized numeric vector}
    \item{\code{x370}}{an anonymized numeric vector}
    \item{\code{x371}}{an anonymized numeric vector}
    \item{\code{x372}}{an anonymized numeric vector}
    \item{\code{x373}}{an anonymized numeric vector}
    \item{\code{x374}}{an anonymized numeric vector}
    \item{\code{x375}}{an anonymized numeric vector}
    \item{\code{x376}}{an anonymized numeric vector}
    \item{\code{x377}}{an anonymized numeric vector}
    \item{\code{x378}}{an anonymized numeric vector}
    \item{\code{x379}}{an anonymized numeric vector}
    \item{\code{x380}}{an anonymized numeric vector}
    \item{\code{x381}}{an anonymized numeric vector}
    \item{\code{x382}}{an anonymized numeric vector}
    \item{\code{x383}}{an anonymized numeric vector}
    \item{\code{x384}}{an anonymized numeric vector}
    \item{\code{x385}}{an anonymized numeric vector}
    \item{\code{x386}}{an anonymized numeric vector}
    \item{\code{x387}}{an anonymized numeric vector}
    \item{\code{x388}}{an anonymized numeric vector}
    \item{\code{x389}}{an anonymized numeric vector}
    \item{\code{x390}}{an anonymized numeric vector}
    \item{\code{x391}}{an anonymized numeric vector}
    \item{\code{x392}}{an anonymized numeric vector}
    \item{\code{x393}}{an anonymized numeric vector}
    \item{\code{x394}}{an anonymized numeric vector}
    \item{\code{x395}}{an anonymized numeric vector}
    \item{\code{x396}}{an anonymized numeric vector}
    \item{\code{x397}}{an anonymized numeric vector}
    \item{\code{x398}}{an anonymized numeric vector}
    \item{\code{x399}}{an anonymized numeric vector}
    \item{\code{x400}}{an anonymized numeric vector}
    \item{\code{x401}}{an anonymized numeric vector}
    \item{\code{x402}}{an anonymized numeric vector}
    \item{\code{x403}}{an anonymized numeric vector}
    \item{\code{x404}}{an anonymized numeric vector}
    \item{\code{x405}}{an anonymized numeric vector}
    \item{\code{x406}}{an anonymized numeric vector}
    \item{\code{x407}}{an anonymized numeric vector}
    \item{\code{x408}}{an anonymized numeric vector}
    \item{\code{x409}}{an anonymized numeric vector}
    \item{\code{x410}}{an anonymized numeric vector}
    \item{\code{x411}}{an anonymized numeric vector}
    \item{\code{x412}}{an anonymized numeric vector}
    \item{\code{x413}}{an anonymized numeric vector}
    \item{\code{x414}}{an anonymized numeric vector}
    \item{\code{x415}}{an anonymized numeric vector}
    \item{\code{x416}}{an anonymized numeric vector}
    \item{\code{x417}}{an anonymized numeric vector}
    \item{\code{x418}}{an anonymized numeric vector}
    \item{\code{x419}}{an anonymized numeric vector}
  }
}
\details{
This example is inspired by the Online Product Sales competition on kaggle.com.  The goal is to isolate the minimum number predictors required for accurately predicting \code{Profit}.  Since the data is based on an actual case, all predictors are anonymized (some were originally categorical but are treated as numerical for the example).
}
\source{
Inspired by \url{https://www.kaggle.com/c/online-sales}
}
\keyword{datasets}
