\name{SMALLFLYER}
\alias{SMALLFLYER}
\docType{data}
\title{
Interest in a frequent flier program (small version)
}
\description{
Interest in a frequent flier program (artificial)
}
\usage{data("SMALLFLYER")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male}}
    \item{\code{Interest}}{a factor with levels \code{No} \code{Yes}}
  }
}
\details{
This artificial datasets tabulates the interest in a new frequent flyer program based on gender.  A larger version of the same data is in \code{LARGEFLYER}.  
}
\keyword{datasets}
