\name{SPEED}
\alias{SPEED}
\docType{data}
\title{
Speed vs. Fuel Efficiency
}
\description{
Speed vs. Fuel Efficiency
}
\usage{data("SPEED")}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{AverageSpeed}}{a numeric vector describing the average speed that the vehicle was driven}
    \item{\code{FuelEfficiency}}{a numeric vector describing the measured fuel efficiency }
  }
}
\details{
The relationship between fuel efficiency and speed is non-monotonic.
}
\source{
Artificial 
}

\keyword{datasets}
