\name{outlier_demo}
\alias{outlier_demo}
\alias{outlier.demo}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Interactive demonstration of the effect of an outlier on a regression
}
\description{
This function shows regression lines on user-defined data before and after adding an additional point.}
\usage{
outlier_demo()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
No arguments
}
\details{
This function allows the user to generate data by click on a plot.  Once two points are added, the least squares regression line is draw.  When an additional point is added, the regression line updates while also showing the line without that point.  The effect of outliers on a regression line can easily be illustrated.  

Note:  To end the demo, you MUST click on the red box labeled "End" (or press Escape, which will return an error)

}

\references{
Introduction to Regression and Modeling
}

\author{
Adam Petrie
}
